+++ cub3d/info_init.c
@@ -6,7 +6,7 @@
 /*   By: hyungjpa <hyungjpa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 12:26:16 by hyungjpa          #+#    #+#             */
-/*   Updated: 2023/08/08 18:26:28 by hyungjpa         ###   ########.fr       */
+/*   Updated: 2023/08/19 02:07:36 by hyungjpa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,55 +92,4 @@ t_flag init_info(t_info *info)
     info->player_x = 0;
     info->player_y = 0;
     return (SUCCESS);
-}
-
-void var_init(t_var *var, t_info *info)
-{
-    var->info = info;
-    var->posX = (double)info->player_x;
-    var->posY = (double)info->player_y;
-
-    // 플레이어가 바라보는 방향 벡터 결정
-    if (info->player_dir == 'N')
-    {
-        var->dirX = 0;
-        var->dirY = -1;
-        var->planeX = -0.66;
-        var->planeY = 0;
-    }
-    if (info->player_dir == 'S')
-    {
-        var->dirX = 0;
-        var->dirY = 1;
-        var->planeX = 0.66;
-        var->planeY = 0;
-    }
-    if (info->player_dir == 'W')
-    {
-        var->dirX = -1;
-        var->dirY = 0;
-        var->planeX = 0;
-        var->planeY = -0.66;
-    }
-    if (info->player_dir == 'E')
-    {
-        var->dirX = 1;
-        var->dirY = 0;
-        var->planeX = 0;
-        var->planeY = 0.66;
-    }
-
-    var->cameraX = 0;
-    var->mapX = 0;
-    var->mapY = 0;
-    var->side = 0;
-    var->rayDirX = 0;
-    var->rayDirY = 0;
-    var->stepX = 0;
-    var->stepY = 0;
-    var->sideDistX = 0;
-    var->sideDistY = 0;
-    var->deltaDistX = 0;
-    var->deltaDistX = 0;
-    var->perpWallDist = 0;
 }
